split WlPointer out of wl_seat.cpp
