+++ src/server/frontend/wayland/wl_pointer.h
@@ -1,0 +1,67 @@
+/*
+ * Copyright © 2018 Canonical Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * Authored by: Christopher James Halse Rogers <christopher.halse.rogers@canonical.com>
+ */
+
+#ifndef WL_POINTER_H
+#define WL_POINTER_H
+
+#include "core_generated_interfaces.h"
+
+struct MirInputEvent;
+typedef unsigned int MirPointerButtons;
+
+namespace mir
+{
+
+class Executor;
+
+namespace frontend
+{
+
+class WlPointer : public wayland::Pointer
+{
+public:
+
+    WlPointer(
+        wl_client* client,
+        wl_resource* parent,
+        uint32_t id,
+        std::function<void(WlPointer*)> const& on_destroy,
+        std::shared_ptr<mir::Executor> const& executor);
+
+    ~WlPointer();
+
+    void handle_event(MirInputEvent const* event, wl_resource* target);
+
+private:
+    wl_display* const display;
+    std::shared_ptr<mir::Executor> const executor;
+
+    std::function<void(WlPointer*)> on_destroy;
+    std::shared_ptr<bool> const destroyed;
+
+    MirPointerButtons last_set{0};
+    float last_x{0}, last_y{0}, last_vscroll{0}, last_hscroll{0};
+
+    void set_cursor(uint32_t serial, std::experimental::optional<wl_resource*> const& surface, int32_t hotspot_x, int32_t hotspot_y) override;
+    void release() override;
+};
+
+}
+}
+
+#endif // WL_POINTER_H
