+++ src/server/frontend/wayland/wl_seat.cpp
@@ -23,6 +23,7 @@
 #include "wayland_utils.h"
 #include "wl_surface.h"
 #include "wl_keybaord.h"
+#include "wl_pointer.h"
 
 #include "mir/executor.h"
 #include "mir/client/event.h"
@@ -34,8 +35,6 @@
 #include "mir/input/keymap.h"
 #include "mir/input/mir_keyboard_config.h"
 
-#include <linux/input-event-codes.h>
-
 #include <mutex>
 #include <unordered_set>
 
@@ -45,173 +44,6 @@ class Executor;
 
 namespace frontend
 {
-class WlPointer;
-class WlTouch;
-
-class WlPointer : public wayland::Pointer
-{
-public:
-
-    WlPointer(
-        wl_client* client,
-        wl_resource* parent,
-        uint32_t id,
-        std::function<void(WlPointer*)> const& on_destroy,
-        std::shared_ptr<mir::Executor> const& executor)
-        : Pointer(client, parent, id),
-          display{wl_client_get_display(client)},
-          executor{executor},
-          on_destroy{on_destroy},
-          destroyed{std::make_shared<bool>(false)}
-    {
-    }
-
-    ~WlPointer()
-    {
-        *destroyed = true;
-        on_destroy(this);
-    }
-
-    void handle_event(MirInputEvent const* event, wl_resource* target)
-    {
-        executor->spawn(run_unless(
-            destroyed,
-            [
-                ev = mir::client::Event{mir_input_event_get_event(event)},
-                target,
-                target_window_destroyed = WlSurface::from(target)->destroyed_flag(),
-                this
-            ]()
-            {
-                if (*target_window_destroyed)
-                    return;
-
-                auto const serial = wl_display_next_serial(display);
-                auto const event = mir_event_get_input_event(ev);
-                auto const pointer_event = mir_input_event_get_pointer_event(event);
-
-                switch(mir_pointer_event_action(pointer_event))
-                {
-                    case mir_pointer_action_button_down:
-                    case mir_pointer_action_button_up:
-                    {
-                        auto const current_set  = mir_pointer_event_buttons(pointer_event);
-                        auto const current_time = mir_input_event_get_event_time(event) / 1000;
-
-                        for (auto const& mapping :
-                            {
-                                std::make_pair(mir_pointer_button_primary, BTN_LEFT),
-                                std::make_pair(mir_pointer_button_secondary, BTN_RIGHT),
-                                std::make_pair(mir_pointer_button_tertiary, BTN_MIDDLE),
-                                std::make_pair(mir_pointer_button_back, BTN_BACK),
-                                std::make_pair(mir_pointer_button_forward, BTN_FORWARD),
-                                std::make_pair(mir_pointer_button_side, BTN_SIDE),
-                                std::make_pair(mir_pointer_button_task, BTN_TASK),
-                                std::make_pair(mir_pointer_button_extra, BTN_EXTRA)
-                            })
-                        {
-                            if (mapping.first & (current_set ^ last_set))
-                            {
-                                auto const action = (mapping.first & current_set) ?
-                                                    WL_POINTER_BUTTON_STATE_PRESSED :
-                                                    WL_POINTER_BUTTON_STATE_RELEASED;
-
-                                wl_pointer_send_button(resource, serial, current_time, mapping.second, action);
-                            }
-                        }
-
-                        last_set = current_set;
-                        break;
-                    }
-                    case mir_pointer_action_enter:
-                    {
-                        wl_pointer_send_enter(
-                            resource,
-                            serial,
-                            target,
-                            wl_fixed_from_double(mir_pointer_event_axis_value(pointer_event, mir_pointer_axis_x)),
-                            wl_fixed_from_double(mir_pointer_event_axis_value(pointer_event, mir_pointer_axis_y)));
-                        break;
-                    }
-                    case mir_pointer_action_leave:
-                    {
-                        wl_pointer_send_leave(
-                            resource,
-                            serial,
-                            target);
-                        break;
-                    }
-                    case mir_pointer_action_motion:
-                    {
-                        auto x = mir_pointer_event_axis_value(pointer_event, mir_pointer_axis_x);
-                        auto y = mir_pointer_event_axis_value(pointer_event, mir_pointer_axis_y);
-                        auto vscroll = mir_pointer_event_axis_value(pointer_event, mir_pointer_axis_vscroll);
-                        auto hscroll = mir_pointer_event_axis_value(pointer_event, mir_pointer_axis_hscroll);
-
-                        if ((x != last_x) || (y != last_y))
-                        {
-                            wl_pointer_send_motion(
-                                resource,
-                                mir_input_event_get_event_time(event) / 1000,
-                                wl_fixed_from_double(x),
-                                wl_fixed_from_double(y));
-
-                            last_x = x;
-                            last_y = y;
-                        }
-                        if (vscroll != last_vscroll)
-                        {
-                            wl_pointer_send_axis(
-                                resource,
-                                mir_input_event_get_event_time(event) / 1000,
-                                WL_POINTER_AXIS_VERTICAL_SCROLL,
-                                wl_fixed_from_double(vscroll));
-                            last_vscroll = vscroll;
-                        }
-                        if (hscroll != last_hscroll)
-                        {
-                            wl_pointer_send_axis(
-                                resource,
-                                mir_input_event_get_event_time(event) / 1000,
-                                WL_POINTER_AXIS_HORIZONTAL_SCROLL,
-                                wl_fixed_from_double(hscroll));
-                            last_hscroll = hscroll;
-                        }
-                        break;
-                    }
-                    case mir_pointer_actions:
-                        break;
-                }
-            }));
-    }
-
-    // Pointer interface
-private:
-    wl_display* const display;
-    std::shared_ptr<mir::Executor> const executor;
-
-    std::function<void(WlPointer*)> on_destroy;
-    std::shared_ptr<bool> const destroyed;
-
-    MirPointerButtons last_set{0};
-    float last_x{0}, last_y{0}, last_vscroll{0}, last_hscroll{0};
-
-    void set_cursor(uint32_t serial, std::experimental::optional<wl_resource*> const& surface, int32_t hotspot_x, int32_t hotspot_y) override;
-    void release() override;
-};
-
-void WlPointer::set_cursor(uint32_t serial, std::experimental::optional<wl_resource*> const& surface, int32_t hotspot_x, int32_t hotspot_y)
-{
-    (void)serial;
-    (void)surface;
-    (void)hotspot_x;
-    (void)hotspot_y;
-}
-
-void WlPointer::release()
-{
-    wl_resource_destroy(resource);
-}
 
 class WlTouch : public wayland::Touch
 {
